+++ src/plugins/managesieve/managesieve.c
@@ -144,7 +144,7 @@ static gboolean sieve_read_chunk_cb(SockInfo *source, GIOCondition condition, gp
 	session_set_timeout(session, session->timeout_interval);
 
 	if (session->read_buf_len == 0) {
-		gint read_len = -1;
+		ssize_t read_len = -1;
 
 		if (session->sock)
 			read_len = sock_read(session->sock, session->read_buf, SESSION_BUFFSIZE - 1);
